Make each file a separate test case
