+++ test/ex_type_test.exs
@@ -10,17 +10,17 @@ defmodule ExTypeTest do
 
   doctest ExType
 
-  test "typespec test cases" do
-    for file <- Path.wildcard("#{__DIR__}/ex_type/**/*_test_case.ex") do
-      result = capture_io(fn -> ExType.check(file) end)
+  for file <- Path.wildcard("#{__DIR__}/type_checks/**/*.ex") do
+    test "type check test case #{file}" do
+      result = capture_io(fn -> ExType.check(unquote(file)) end)
       assert String.contains?(result, [Emoji.one_test_pass()])
       refute String.contains?(result, [Emoji.one_test_fail()])
     end
   end
 
-  test "typespec failure cases" do
-    for file <- Path.wildcard("#{__DIR__}/ex_type/**/*_failure_case.ex") do
-      result = capture_io(fn -> ExType.check(file) end)
+  for file <- Path.wildcard("#{__DIR__}/type_failures/**/*.ex") do
+    test "type check failure case #{file}" do
+      result = capture_io(fn -> ExType.check(unquote(file)) end)
       refute String.contains?(result, [Emoji.one_test_pass()])
       assert String.contains?(result, [Emoji.one_test_fail()])
     end
