+++ test/type_checks/parametric/function.ex
@@ -1,0 +1,6 @@
+defmodule TypeChecks.Parametric.Function do
+  @spec fun(t) :: t when t: var
+  def fun(t) do
+    t
+  end
+end
