+++ test/type_checks/simple/function_without_arg.ex
@@ -1,0 +1,6 @@
+defmodule TypeChecks.Simple.FunctionWithoutArg do
+  @spec fun() :: integer()
+  def fun do
+    23
+  end
+end
