+++ test/type_failures/atoms/case.ex
@@ -1,4 +1,4 @@
-defmodule AtomFailureCase do
+defmodule TypeFailures.Atoms.Case do
   @spec wrong_atom(integer() | :foo) :: :bar
   def wrong_atom(x) do
     case x do
