+++ test/type_failures/atoms/one_argument.ex
@@ -1,0 +1,6 @@
+defmodule TypeFailures.Atoms.OneArgument do
+  @spec wrong_atom(any()) :: :foo
+  def wrong_atom(_) do
+    :bar
+  end
+end
